fix invalid page tokens for entities without ids
entities such as `Time` and `State` don't have IDs, so they're given a dummy ID which happens to be significantly shorter than normal. this means the page token is shorter than the expected 16 byte minimum, but is otherwise valid

(please test this. i made this pr on my phone with the GitHub web editor)